Pass `different_domains` to compare
Bug fix: we cannot print the information on domains of definition
before the call to compare because then it appears before the tag, as
if it concerned the previous variable. So we have to store the
information in variable `different_domains` and pass this variable to
compare. This revision requires an updated revision (b2bed75) of
library Jumble.