Add install rules in `CMakeLists.txt`
Add install rules in `CMakeLists.txt`. We use the option COMPONENT
runtime so we can install only the executable files, not the
libraries, when FETCH is ON.  Rename `max_diff_nc.sh` to
`max_diff_nc_in.sh`, which is input to `configure_file` in
`CMakeLists.txt`, for automatic replacement of path to Fortran
executable.